scaler

+++ gcn.py
@@ -128,7 +128,7 @@ def main(model_type, dataset_dir, csv_filename, dataset_type, dataset_id, layer,
 				optimizer.zero_grad()
 
 				# forward + backward + optimize
-				outputs = net(inputs)
+				outputs = net(inputs.float())
 				loss = criterion(outputs, labels)
 				loss.backward()
 				optimizer.step()
